Generate and render some agents
Introduce a sprite prefab for Agents, and give the Agent code
appropriate getters/setters. Use these to manage Agent state
and traits.

Additionally, change the unnecessary ints into floats, such
as world width, structure sizes/coordinates, etc.

Move the structures back in the scene by 5 units to allow the
Agent positions (as Vector2s) to default to being in front of
them.

Generate a random valid position for each agent, where 'valid'
is defined as 'not inside a structure'. A hack-ish approach to
this is included, where invalid positions are re-randomized
until success. Considering 20-40% of the map is available in
any given run, this is unlikely to happen more than a few times.