Introduce first stab at agent percepts
Goal is to bee-line towards some entertaining behavior,
so although this solution is weak, it should get part of
the way there.

Percepts are just a struct-y class for now, containing very
basic information each agent may be interested in.

This includes walls and other agents, for now.

This also introduces the piping to get those percepts to
the agent in its time of greatest need, i.e., when it's
updating its plan.