Add fall-through behavior functionality
Behaviors can now be stacked. Each behavior checks to see if
the circumstances (in the percepts and agent state) should be
responded to. If so, that behavior then updates its plan and
returns true, signaling to the FallThroughBehavior (or, indeed,
any parent behavior) that it has made a choice. That then becomes
the parent behavior's plan.

With this in place, the 'random walk' behavior baked into both
FleeBehavior and PursueBehavior has been factored out and added to
the agent spawning method.