Introduce ActionArbiter class to handle agent-agent actions
ActionArbiter resolves actions that are potentially opposed,
or that might have become impossible since the agent wanted to attempt
them. I.e., if a zombie attempts to convert a human because the zombie
was briefly close enough to try, but in the time since then the human
fled out of range, the convert should fail.

Eventually this will allow for directly opposing actions - e.g. a human
fighting off a zombie before being converted.

In preparation of the Convert (zombie->human) action, this commit also
introduces a 'configureAs' method to the Agent class. This is now where
agent type defaults (zombie, human, corpse, etc.) are kept. This allows
for easy conversion between different types, with or without baselining
the agent back to a default state.