Fix bug causing agents to move right at startup
The primary cause of this was WanderBehavior not replotting
until a certain amount of time had passed. Initializing it with
a random direction now.

Additionally, made it so that all behaviors would be able to
trigger in the first cycle (according to their 'update time' param)
by setting the initial value to negative.

Also added in a 'first cycle plan' step that prevents actions
from executing until the first plan cycle has happened.